requirements.txt
Flask
Flask-WTF
gunicorn
python-dotenv
PyPDF2
python-docx
openai
